move fetched records controller up a level
